stripping down gql_utils and putting analytics functions in diff file
